yuzu_cmd: Split emu_window OpenGL implementation into its own file
